Переименовал Класс WindowAuthorization
- Переименовал Класс WindowAuthorization в Класс Authorization
- Добавил Класс Modal В который вынес модальное окно из WindowAuthorization
