--- src/scss/partials/_window-authorization.scss
@@ -1,38 +1,0 @@
-.window-authorization {
-	padding: 0 5px;
-
-	label {
-		display: block;
-		margin: 5px 0;
-	}
-
-	input {
-		outline: none;
-		border: 1px solid black;
-		border-radius: 5px;
-		padding: 4px;
-		font-family: inherit;
-		width: 100%;
-	}
-
-	&-button {
-		display: flex;
-		justify-content: center;
-		margin: 10px 0;
-		font-family: inherit;
-		font-size: 20px;
-
-		&__item {
-			@include ButtonBG();
-			margin-right: 10px;
-
-			&:last-child {
-				margin-right: 0;
-			}
-
-			&:active {
-				@include ButtonClickBG();
-			}
-		}
-	}
-}
