+++ src/scss/style.scss
@@ -4,9 +4,9 @@
 @import "globals/mixins";
 @import "globals/steyle";
 
-@import "partials/modal-overlay";
-@import "partials/modal";
 @import "partials/header";
 @import "partials/main";
 @import "partials/serial-card";
-@import "partials/window-authorization";
+@import "partials/modal-overlay";
+@import "partials/modal";
+@import "partials/modal-authorization";
